requirements.txt
# RAAF - Resume Assessment Automation Framework
# Python dependencies

# PDF Processing
pdfplumber>=0.10.0
# Alternative: pymupdf>=1.23.0

# DOCX Processing
python-docx>=1.1.0

# Configuration
PyYAML>=6.0

# Data Processing
pandas>=2.0.0

# HTTP Requests (for PCR API)
requests>=2.31.0

# Development/Testing
pytest>=7.4.0
black>=24.0.0
